move the InputFiles map construction from DataIterator to RowDataIterator so that it is constructed per FileScanTask (not per CombinedScanTask)
